requirements.txt
asgiref==3.5.2
Django==4.1.3
Pillow==9.3.0
sqlparse==0.4.3
tzdata==2022.7
django-filter~=22.1